editor: move document model in own file
